fuzz: Don't canonicalize NaN while fuzzing
This came up over the weekend where the spec interpreter and Wasmtime
were differing on the result of a wasm program, but the difference came
down to how nan bits are represented. Because Wasmtime was forcibly
canonicalizing nan representations it created a mismatch in the results
of the interpreter and not. While permissible by the spec this hinders
fuzzing, so this commit attempts to address this to ensure that the spec
interpreter produces the same NaN values as Wasmtime does (and vice
versa).

NaN canonicalization was enabled for fuzzing in #1334 which tracks back
to #1332 and a desire for more determinism when fuzzing. I believe,
though, that NaN bits are always deterministic within one machine,
they're just "nondeterministic" across machines (although I think in
practice all cpus produce the same nan bits?) In any case I don't think
that canonicalization is required for the fuzzers and disabling it helps
us align with what the spec interpreter is doing as well.