Don't parse symbol names allocating a `CodeMemory`
One of the slightly-more-expensive parts of creating a `CodeMemory` at
this time is how symbol names in the object are parsed to determine
whether they're for a defined function or a trampoline. Given the rigid
structure of the object file, however, we don't actually need to do
parsing and instead we can guess the `SymbolIndex` associated with each
desired function. This means we can eschew parsing altogether and we
never actually need to guess the name of the symbol.

This commit also defers allocation of trampoline/function maps out of
`CodeMemory` into  the instantiation portions since that's already
rebuilding the maps and there's no need to build them twice.