Allow for opting out of uffd even if the uffd feature is enabled.
This commit adds a `PoolingPageFaultStrategy` enum to the Wasmtime API to
control whether the operating system or Wasmtime will handle page faults in
linear memory.

The latter requires the `uffd` feature and Linux as the target OS.

With this change, a pooling allocator can be created that uses the operating
system's page fault handling even on Linux with the `uffd` feature enabled.